Add files via upload
Python is a multi-purpose language and is widely used for scripting. Its usage is not just limited to solving complex calculations but also to automating daily life tasks. Let’s say we want to track any Amazon product availability and grab the deal when the product availability changes and inform the user of availability through email. It will be great fun to write a Python script for this.